refactor: move ProcessDetector to node; export noopDetector for Browser
