+++ packages/opentelemetry-resources/src/detectors/index.ts
@@ -16,7 +16,5 @@
 
 export * from './BrowserDetector';
 export * from './EnvDetector';
-export * from './ProcessDetector';
 export * from './BrowserDetectorSync';
 export * from './EnvDetectorSync';
-export * from './ProcessDetectorSync';
