+++ packages/opentelemetry-resources/src/platform/browser/ProcessDetectorSync.ts
@@ -14,10 +14,6 @@
  * limitations under the License.
  */
 
-export * from './default-service-name';
-export * from './HostDetector';
-export * from './OSDetector';
-export * from './HostDetectorSync';
-export * from './OSDetectorSync';
-export * from './ProcessDetector';
-export * from './ProcessDetectorSync';
+import { noopDetector } from '../../detectors/NoopDetector';
+
+export const processDetectorSync = noopDetector;
